# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Foods",
        "version": "0.1"
    },
    "paths": {
        "/ingredients/all": {
            "get": {
                "tags": [
                    "Ingredients"
                ],
                "description": "Retrieve all ingredients",
                "operationId": "App\\Http\\Controllers\\IngredientController::retrieve",
                "responses": {
                    "200": {
                        "description": "Return ingredients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ingredients": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/IngredientDto"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/recipes/all": {
            "get": {
                "tags": [
                    "Recipies"
                ],
                "description": "Retrieve all recepies",
                "operationId": "App\\Http\\Controllers\\RecipeController::retrieve",
                "responses": {
                    "200": {
                        "description": "Return recipes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "recipes": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RecipeDto"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/recipe/add": {
            "post": {
                "tags": [
                    "Recipies"
                ],
                "description": "Add a recipe",
                "operationId": "App\\Http\\Controllers\\RecipeController::add",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "ingredients",
                                    "preparationTime",
                                    "process"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "name"
                                    },
                                    "ingredients": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "unite/gramme/millimeter"
                                                },
                                                "qty": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "preparationTime": {
                                        "type": "integer",
                                        "example": "300"
                                    },
                                    "process": {
                                        "type": "string",
                                        "example": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return the recipeId (recipe can be acceeded by url /recipe/{id}"
                    },
                    "401": {
                        "description": "User unauthorize"
                    },
                    "400": {
                        "description": "Bad request or server error during saving"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/user/inscription": {
            "post": {
                "tags": [
                    "Register"
                ],
                "description": "Register by email, password, user name, user firstname",
                "operationId": "App\\Http\\Controllers\\UserController::inscription",
                "requestBody": {
                    "description": "Pass user informations",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "userEmail",
                                    "userPassword",
                                    "userName",
                                    "userFirstName"
                                ],
                                "properties": {
                                    "userEmail": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "userPassword": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "userName": {
                                        "type": "string",
                                        "example": "Lambert"
                                    },
                                    "userFirstName": {
                                        "type": "string",
                                        "example": "Dylan"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Registration failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Email already use"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Registration complete"
                    }
                }
            }
        },
        "/user/connexion": {
            "post": {
                "tags": [
                    "Connexion"
                ],
                "description": "Retrieve user by email, password",
                "operationId": "App\\Http\\Controllers\\UserController::connexion",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "userEmail",
                                    "userPassword"
                                ],
                                "properties": {
                                    "userEmail": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "userPassword": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return connexion token, need to be placed on all request (in header bearer style)"
                    },
                    "400": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "user not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "IngredientDto": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Boeuf Bourgignon"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "RecipeDto"
                }
            },
            "RecipeDto": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Boeuf Bourgignon"
                    },
                    "preparationTime": {
                        "type": "string",
                        "example": "16m"
                    },
                    "ingredients": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "ingredientName": {
                                    "type": "string",
                                    "example": "Carrotes"
                                },
                                "quantity": {
                                    "type": "string",
                                    "example": "6 Kg"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "recipe": {
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "RecipeDto"
                }
            },
            "UserDto": {
                "properties": {
                    "userId": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": "true"
                    },
                    "userEmail": {
                        "description": "User unique email address",
                        "type": "string",
                        "format": "email",
                        "readOnly": "true",
                        "example": "user@gmail.com"
                    },
                    "userName": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "Doe"
                    },
                    "userFirstName": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "John"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserDto"
                }
            }
        }
    },
    "security": [
        []
    ]
}